Update adding new well measurements functionality to use the new DB naming
